hello test

+++ lz77_encoder.cpp
@@ -25,7 +25,7 @@ void lz77_encoder::open_output_file(char *output_file) {
 
 void lz77_encoder::read_block() {
     uint size = input.readsome(reinterpret_cast<char *>(block_buffer), BUFFER_SIZE);
-    for (int i = 0; i < size; i++) {
+    for (uint i = 0; i < size; i++) {
         buffer.enqueue(block_buffer[i]);
     }
 }
@@ -136,7 +136,7 @@ std::pair<ushort, ushort> lz77_encoder::findMatching() {
 }
 
 void lz77_encoder::shift_buffer(uint shift) {
-    if (buffer.size() < shift) {
+    if ((uint)buffer.size() < shift) {
         read_block();
     }
